monorepo-test-runner: run all tests if a config file is changed
Changes in .jest.config.js will trigger all tests to run.
The test runner will also load the config file and check if any
files in setupFilesAfterEnv or setupFiles has changed. If so
all tests will be executed.

Note that it does not check files in project.setupFiles. If you
want changes to a file to trigger all tests, it should be specified
at the root level of the jest configuration.

adeira-source-id: adef538333367362e670f1df4a2bf8335a831b11